pyproject.toml
[tool.poetry]
name = "ninjad"
version = "0.1.0"
description = "Ninja Template"
authors = ["Samuel Ajisegiri <samajisegiri@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.9.2"
django-configurations = "^2.5.1"
django-ninja = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
